Add install-method config support and drop symlinks
Add support for install-method config which can be set to 'classic'
or 'strict'. The root directory where setup dirs, templates, and
copyfiles are created is based on the value of install-method. If
install-method=classic (default) the root directory is / and if
install-method=strict the root directory is $SNAP_COMMON.

Additionally the symlinks setup key is dropped. Classic snaps will
install files into / as defined by install-method=classic.

Change-Id: I6750957654e75169108e6e866ab0d43e588cbd12