Add install setup support and drop symlinks
Add support for an install setup key that is set in snap-openstack.yaml
and defines whether the snap should be installed as a 'classic' or
'strict' snap. This determines the root directory where setup dirs,
templates, and copyfiles are created. If install is set to classic
the root directory will be / and if install is set to strict the root
directory will be $SNAP_COMMON.

Additionally the symlinks setup key is dropped. Classic snaps will
install files into / as defined by the install key.

Change-Id: I6750957654e75169108e6e866ab0d43e588cbd12