Enable dynamic creation of uwsgi pyargv option
Specifying hard-coded config files via the uwsgi pyargv option
caused failures when a file didn't exist.

This patch enables dynamic creation of the pyargv option. This
value is pushed onto the context for templated file generation
which is now supported as a subelement of an entry point as
well as a top level global option.

Change-Id: I1d0fae42b2e43fe8808fda3de83e122502233a4c