Always execute copyfiles and template directives
Targets for copyfiles and template directives should be considered
as owned by the snap, and as a result should always be executed,
irrespective of whether the target file exists or not.

This ensures that any use of {{ snap }} is current, and that changes
to templates in the snap are reflected on refresh.

Execute copyfiles prior to template rendering to ensure templated
files overwrite any files copied as a copyfiles target.

Change-Id: I97bf668cc86099227670a00defd4ee3c36b58ed7