Allow multiple OS configs and flexible overrides
The config override patch was a bit too aggressive in not allowing
multiple OpenStack config files/dirs to be specified. Some projects
need to specify more than one, so this re-enables the ability to
specify multiple OpenStack config files/dirs.

Additionally this makes the override of OpenStack config files more
flexible in that individual files can be overridden while others are
not.

Change-Id: Id512a96e3a7b408b519f29f860d522a584f1e5d9